+++ packages/Chem/src/demo/demo.ts
@@ -11,7 +11,6 @@ import {CLIFFS_DF_NAME, activityCliffsIdx} from '@datagrok-libraries/ml/src/view
 import {BitArrayMetricsNames} from '@datagrok-libraries/ml/src/typed-metrics';
 import {DimReductionMethods} from '@datagrok-libraries/ml/src/multi-column-dimensionality-reduction/types';
 import {ScaffoldTreeViewer} from '../widgets/scaffold-tree';
-import {MatchedMolecularPairsViewer} from '../analysis/molecular-matched-pairs/mmp-viewer/mmp-viewer';
 
 
 export async function _demoChemOverview(): Promise<void> {
@@ -195,26 +194,17 @@ export async function _demoMMPA(): Promise<void> {
     const layout = DG.ViewLayout.fromJson(layoutString);
     tv.loadLayout(layout);
     tv.dataFrame.currentRowIdx = 0;
-    let mmpViewer: MatchedMolecularPairsViewer | null = null;
     try {
       await awaitCheck(() => {
         for (const v of tv.viewers) {
-          console.log(v.type);
-          if (v.type === 'Matched Molecular Pairs Analysis')
+          if (v.type === 'Matched Molecular Pairs Analysis') {
+            v.helpUrl = 'https://raw.githubusercontent.com/datagrok-ai/public/refs/heads/master/help/datagrok/solutions/domains/chem/chem.md#matched-molecular-pairs';
             return true;
+          }
         }
         return false;
       }, '', 20000);
     } catch (e) {};
-    for (const v of tv.viewers) {
-      if (v.type === 'Matched Molecular Pairs Analysis') {
-        mmpViewer = v as MatchedMolecularPairsViewer;
-        break;
-      }
-    }
-    if (mmpViewer == null)
-      return;
-    mmpViewer.helpUrl = 'https://raw.githubusercontent.com/datagrok-ai/public/refs/heads/master/help/datagrok/solutions/domains/chem/chem.md#matched-molecular-pairs';
     setTimeout(()=> {
       grok.shell.windows.showHelp = true;
       grok.shell.windows.help.showHelp('/help/datagrok/solutions/domains/chem/chem#matched-molecular-pairs');